feat(manifest-beta.json): add manifest file for beta version of the Steemit plugin
The manifest-beta.json file is added to the project. This file contains the metadata for the beta version of the Steemit plugin. It includes the plugin ID, name, version, minimum app version required, description, author information, desktop-only flag, and funding URL. This manifest file will be used to configure and publish the beta version of the plugin.